requirements.txt
# LlamaIndex core e componentes
llama-index==0.14.7
llama-index-core==0.14.7
llama-index-embeddings-google-genai==0.3.1
llama-index-retrievers-bm25==0.3.0
google-genai==1.47.0

# Dependências para processamento de texto em português
nltk==3.9.2
pandas==2.2.3
numpy==2.3.4
python-dotenv==1.2.1
unidecode==1.4.0
rank-bm25==0.2.2

# Dependências opcionais para melhor performance
faiss-cpu==1.12.0  # Para busca vetorial mais eficiente (opcional)

# Para baixar o dataset com script utils/download_juris_tcu.py
huggingface-hub==1.0.1
